extract eas-client.ts
